feat(37-05): refactor pool, queue, router for control/data plane separation
- Move execution functions from pool.rs to data_plane.rs
- Worker threads now call data_plane::execute_with_context_manager
- Add ControlPlane integration to WorkQueue for request validation
- Update router to validate requests through control plane before dispatch
- Fix worker_count types (u32 -> usize) across WorkerPool, WorkerHandle, trait
- Fix http/types.rs set_worker_id signature for type consistency
- Fix v8/isolate.rs typo in set_heap_limits logging
- Fix runtime/crypto/ecdsa.rs for p256/p384 pkcs8 API compatibility

+++ src/http/types.rs
@@ -125,7 +125,7 @@ pub struct NanoResponse {
     headers: NanoHeaders,
     body: Option<Bytes>,
     /// Worker ID that processed this request (for logging/debugging)
-    worker_id: Option<usize>,
+    worker_id: Option<u32>,
     /// Isolate ID that processed this request (for logging/debugging)
     isolate_id: Option<String>,
 }
@@ -148,12 +148,12 @@ impl NanoResponse {
     }
 
     /// Get the worker ID that processed this request
-    pub fn worker_id(&self) -> Option<usize> {
+    pub fn worker_id(&self) -> Option<u32> {
         self.worker_id
     }
 
     /// Set the worker ID that processed this request
-    pub fn set_worker_id(&mut self, worker_id: usize) {
+    pub fn set_worker_id(&mut self, worker_id: u32) {
         self.worker_id = Some(worker_id);
     }
